+++ tests/services/test_exchange_rate_service.py
@@ -6,7 +6,7 @@ import pytest
 
 from app.models.currency_pair import CurrencyPair
 from app.models.exchange_rate import ExchangeRate
-from app.models.exchange_rate_service import ExchangeRateService
+from app.services.exchange_rate_service import ExchangeRateService
 from tests.support.database_helper import DatabaseTestHelper
 from tests.support.decimal import quantize_decimal
 from tests.support.factories import build_exchange_rate
